fix employee manager merge issue. add store manager and order model

+++ src/Domain.Model/Factory/AttritionStatistic.cs
@@ -9,7 +9,7 @@
 
         public double PriceOfUnit { get; set; }
 
-        public long AttritionId { get; set; }
+        public int AttritionId { get; set; }
 
         public virtual Attrition Attrition { get; set; }
     }
