+++ src/Domain.Model/Factory/WorkTimeStatistic.cs
@@ -1,8 +1,7 @@
-﻿
-namespace Domain.Model.Factory
+﻿namespace Domain.Model.Factory
 {
-    using System;
     using Infrastructure.Domain.Model;
+    using System;
 
     public class WorkTimeStatistic : Statistic, IAggregateRoot
     {
@@ -10,7 +9,7 @@ namespace Domain.Model.Factory
 
         public double WorkTimeHr { get; set; }
 
-        public long EmployeeId { get; set; }
+        public int EmployeeId { get; set; }
 
         public virtual Employee Employee { get; set; }
     }
